add to producer item edit and delete

+++ src/main/java/com/example/springsecurityexam/repository/MemberRepository.java
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Member save(Member member);
 
-    Member findById(int id);
+    Member findById(long id);
 
     List<Member> findByAuth(RoleType auth);
